Fixed clang tidy issue.

+++ src/fheroes2/battle/battle_action.cpp
@@ -805,9 +805,9 @@ Battle::TargetsInfo Battle::Arena::TargetsForChainLightning( const HeroBase * he
 
     TargetsInfo targets;
 
-    const int32_t firstUnitResist = unit->GetMagicResist( Spell::CHAINLIGHTNING, 0, hero );
+    const uint32_t firstUnitResist = unit->GetMagicResist( Spell::CHAINLIGHTNING, 0, hero );
 
-    if ( firstUnitResist >= static_cast<int32_t>( _randomGenerator.Get( 1, 100 ) ) ) {
+    if ( firstUnitResist >= _randomGenerator.Get( 1, 100 ) ) {
         targets.emplace_back();
         TargetInfo & res = targets.back();
         res.defender = unit;
